test: get last test working
