--- src/taxonomy/import-tags/__mocks__/index.js
@@ -1,1 +1,0 @@
-export { default as taxonomyImportMock } from './taxonomyImportMock'; // eslint-disable-line import/prefer-default-export
