--- src/taxonomy/import-tags/__mocks__/taxonomyImportMock.js
@@ -1,5 +1,0 @@
-export default {
-  name: 'Taxonomy name',
-  exportId: 'taxonomy_export_id',
-  description: 'Taxonomy description',
-};
